filesize (bits) in a 32 bit int will overflow after 250MB, screwing up 2pass ratecontrol. (patch by Loren Merritt <lorenm at u dot washington dot edu>)
git-svn-id: svn://svn.videolan.org/x264/trunk@52 df754926-b1dd-0310-bc7b-ec298dee348c

+++ core/common.h
@@ -359,7 +359,7 @@ struct x264_t
 
         /* per slice info */
         int   i_slice_count[5];
-        int   i_slice_size[5];
+        int64_t i_slice_size[5];
         /* */
         int64_t i_sqe_global[5];
         float   f_psnr_average[5];
@@ -367,7 +367,7 @@ struct x264_t
         float   f_psnr_mean_u[5];
         float   f_psnr_mean_v[5];
         /* */
-        int   i_mb_count[5][18];
+        int64_t i_mb_count[5][18];
 
     } stat;
